requirements.txt
requests
beautifulsoup4
flask
gunicorn
python-dotenv